% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_CgNBFA_approx.R
\name{L_CgNBFA_approx}
\alias{L_CgNBFA_approx}
\title{Compute approximate copula likelihood with negative binomial marginals}
\usage{
L_CgNBFA_approx(obs, size, mu, Gamma, logl = T)
}
\arguments{
\item{obs}{vector. The data point.}

\item{size}{vector. The estimated size parameter.}

\item{mu}{vector. The estimated mean parameter.}

\item{Gamma}{matrix. The estimated correlation matrix.}

\item{logl}{logical. Should the output be log-likelihood?}
}
\value{
The (log-)likelihood.
}
\description{
This function computes the approximated likelihood of a Gaussian copula with
NB marginals. Integration in the calculation of the exact
likelihood can be computationally expensive, especially in high dimensions,
therefore this function provides us with an approximation of the integral.
}
