% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_CgNBFA.R
\name{pred_CgNBFA}
\alias{pred_CgNBFA}
\title{Predict from CgNBFA model}
\usage{
pred_CgNBFA(tr_fa, gp_test, test)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_CgNBFA}.}

\item{gp_test}{vector. Groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{lpd}{A numeric vector of log predictive densities for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the CgNBFA
model \code{train_CgNBFA}.
}
\seealso{
\code{\link{train_CgNBFA}}
}
