% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_CgPFA.R
\name{pred_CgPFA}
\alias{pred_CgPFA}
\title{Predict from CgPFA model}
\usage{
pred_CgPFA(tr_fa, gp_test, test, logl = T)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_CgPFA}.}

\item{gp_test}{vector. Groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{likelihoods}{A numeric vector of (log-)likelihoods for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the CgPFA
model \code{train_CgPFA}.
}
\seealso{
\code{\link{train_CgPFA}}
}
