% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_gNBFA.R
\name{pred_gNBFA}
\alias{pred_gNBFA}
\title{Predict from gNBFA model}
\usage{
pred_gNBFA(tr_fa, gp_test, test, logl = T)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_gNBFA}.}

\item{gp_test}{vector. Groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{likelihoods}{A numeric vector of (log-)likelihoods for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the gNBFA
model \code{train_gNBFA}.
}
\seealso{
\code{\link{train_gNBFA}}
}
