% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_gNBGPFA_LP.R
\name{pred_gNBGPFA_LP}
\alias{pred_gNBGPFA_LP}
\title{Evaluates a negative binomial Gaussian process factor analysis model}
\usage{
pred_gNBGPFA_LP(
  mod,
  X_test,
  ts_test,
  gp_test,
  ts_train,
  gp_train,
  transform_t,
  period_length
)
}
\arguments{
\item{mod}{list. Output of \code{train_gNBGPFA_LP}.}

\item{X_test}{data frame. The test data set, consisting of counts. Columns represent variables, rows represent observations.}

\item{ts_test}{vector. Time-points of observations in X_test.}

\item{gp_test}{vector. Groups for each observation in the test set.}

\item{ts_train}{vector. Time-points of observations in X_train.}

\item{gp_train}{vector. Groups for each observation in the train set.}

\item{transform_t}{function. Function for the transformation of the time-series (for example if we want it on a specific interval).}

\item{period_length}{numeric. Length of the period.}
}
\value{
A list.
\item{data}{Generated quantities for test data.}
\item{log_scores}{Point-wise log scores.}
}
\description{
This function evaluates a negative binomial Gaussian process factor analysis model on test data. Additionally,
it provides generated quantities for the test data.
}
