% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{toy_data}
\alias{toy_data}
\title{toy data}
\format{A list with four elements.
  \describe{
    \item{fa_data}{A data frame with 2000 rows and 8 columns.}
    \item{group}{A vector of group IDs for each observation.}
    \item{splits}{A data frame with three columns which represent
    data splits. 2 stands for the train
    observations and 1 stands for test observations.}
    \item{sup_data}{NULL.}
  }}
\description{
A generated data set with low and overdispersed counts.
}
