% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_E_data.R
\name{toy_E}
\alias{toy_E}
\title{toy E data}
\format{A list with four elements.
  \describe{
    \item{df}{A data frame with 400 rows and 8 columns.}
    \item{ts}{A vector of length 400 of time-points corresponding to each observation.}
    \item{gp}{A vector of length 400 of groups for each observation.}
    \item{splits}{A data frame with 400 rows and two columns. Indexing of a 2-fold CV split of the data.}
  }}
\description{
A generated data set with moderate overdispersed counts (negative binomial) and temporal latent 2-factor structure. No residual covariance.
}
