% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_gMVNGPFA_LP.R
\name{train_gMVNGPFA_LP}
\alias{train_gMVNGPFA_LP}
\title{Train a Gaussian process factor analysis model with full covariance}
\usage{
train_gMVNGPFA_LP(
  X,
  ts,
  gp,
  nfac,
  nit,
  nchain,
  period_length,
  prior_r2,
  prior_r3,
  transform_t,
  tseed = NULL,
  ...
)
}
\arguments{
\item{X}{data frame. The training data set, consisting of counts. Columns represent variables, rows represent observations.}

\item{ts}{vector. Time-points of observations in X.}

\item{gp}{vector. Groups for each observation.}

\item{nfac}{numeric. Number of factors.}

\item{nit}{numeric. Number of iterations.}

\item{nchain}{numeric. Number of chains.}

\item{period_length}{numeric. Length of the period.}

\item{prior_r2}{vector. Parameters (2) for the inverse-gamma prior distribution for the lengthscale of the squared exponential
part of the locally periodic kernel.}

\item{prior_r3}{vector. Parameters (2) for the inverse-gamma prior distribution for the lengthscale of the periodic
part of the locally periodic kernel.}

\item{transform_t}{function. Function for the transformation of the time-series (for example if we want it on a specific interval).}

\item{tseed}{numeric. Seed for sampling Defaults to 1.}
}
\value{
A list.
\item{data}{Data that goes into sampling.}
\item{samps}{An object of S4 class \code{stanfit}. Samples.}
\item{gmap}{Mapping of groups to indexes.}
\item{orig_X}{Original X.}
}
\description{
This function trains a Gaussian process factor analysis model with full covariance and a locally periodic kernel in Stan. Uses a normal likelihood and standardized data.
}
