% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_gNBGPFAC_LP.R
\name{train_gNBGPFAC}
\alias{train_gNBGPFAC}
\title{Train a negative binomial Gaussian process factor analysis with copula}
\usage{
train_gNBGPFAC(
  X,
  ts,
  gp,
  nfac,
  nit,
  nchain,
  period_length,
  prior_r2,
  prior_r3,
  prior_phi_type = 1,
  prior_phi,
  transform_t,
  tseed = NULL,
  init_list = NULL,
  ...
)
}
\arguments{
\item{X}{data frame. The training data set, consisting of counts. Columns represent variables, rows represent observations.}

\item{ts}{vector. Time-points of observations in X.}

\item{gp}{vector. Groups for each observation.}

\item{nfac}{numeric. Number of factors.}

\item{nit}{numeric. Number of iterations.}

\item{nchain}{numeric. Number of chains.}

\item{period_length}{numeric. Length of the period.}

\item{prior_r2}{vector. Parameters (2) for the inverse-gamma prior distribution for the lengthscale of the squared exponential
part of the locally periodic kernel.}

\item{prior_r3}{vector. Parameters (2) for the inverse-gamma prior distribution for the lengthscale of the periodic
part of the locally periodic kernel.}

\item{prior_phi_type}{numeric. If 1, then it uses a gamma prior for the overdispersion parameter. If 2 it uses an inverse-gamma
prior.}

\item{prior_phi}{numeric. Parameters (2) of the gamma or inverse-gamma distribution for the prior of overdispersion parameter phi.}

\item{transform_t}{function. Function for the transformation of the time-series (for example if we want it on a specific interval).}

\item{tseed}{numeric. Seed for sampling Defaults to 1.}

\item{init_list}{list. List of initialization parameters for the model. We recommend using the posterior means of gNBGPFA model.}
}
\value{
A list.
\item{data}{Data that goes into sampling.}
\item{samps}{An object of S4 class \code{stanfit}. Samples.}
\item{gmap}{Mapping of groups to indexes.}
}
\description{
This function trains a negative binomial Gaussian process factor analysis with copula with a locally periodic kernel in Stan.
}
