% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_gPFA.R
\name{pred_gPFA}
\alias{pred_gPFA}
\title{Predict from gPFA model}
\usage{
pred_gPFA(tr_fa, gp_test, test, logl = T)
}
\arguments{
\item{tr_fa}{An object resulting from \code{train_gPFA}.}

\item{gp_test}{vector. Groups for each observation.}

\item{test}{data frame. Test data set for evaluation.}

\item{logl}{logical. Return log-likelihood? Defaults to TRUE.}
}
\value{
A list.
  \item{predictions}{A numeric vector of predictions for all observations
  in the test set.}
  \item{likelihoods}{A numeric vector of (log-)likelihoods for all
  observations in the test set.}
}
\description{
This function predicts based on the estimation of the gPFA
model \code{train_gPFA}.
}
\seealso{
\code{\link{train_gPFA}}
}
